+++ include/sim_data_loader.h
@@ -28,20 +28,20 @@ class SimDataLoader {
   typedef std::shared_ptr<SimDataLoader> Ptr;
   SimDataLoader();
   void LoadDataSet(const std::string& dataset_path);
-  bool PopOutMea(ReplaySensorType& type, double& ts);
-  std::vector<SemanticMea::Ptr> GetSemanticMeas(const double timestamp);
-  std::vector<KinematicMea::Ptr> GetKinematicMeas(const double timestamp);
+  bool PopOutMea(ReplaySensorType& type, long long& ts);
+  std::vector<SemanticMea::Ptr> GetSemanticMeas(const long long timestamp);
+  std::vector<KinematicMea::Ptr> GetKinematicMeas(const long long timestamp);
 
  private:
   void load_dataset_pose(const std::string& pose_file);
   void load_dataset_map(const std::string& map_file);
   void generate_semantic_meas();
 
-  std::map<double, Pose> _gt_pose;
-  std::map<double, std::vector<SemanticMea::Ptr>> _semantic_mea;
-  std::map<double, std::vector<KinematicMea::Ptr>> _kinematic_mea;
-  std::map<double, std::vector<ReplaySensorType>> _mea_seq;
-  std::map<double, std::vector<ReplaySensorType>>::iterator _mea_it;
+  std::map<long long, Pose> _gt_pose;
+  std::map<long long, std::vector<SemanticMea::Ptr>> _semantic_mea;
+  std::map<long long, std::vector<KinematicMea::Ptr>> _kinematic_mea;
+  std::map<long long, std::vector<ReplaySensorType>> _mea_seq;
+  std::map<long long, std::vector<ReplaySensorType>>::iterator _mea_it;
   int _mea_id{0};
 };
 }  // namespace apa_slam
